+++ example/next.config.js
@@ -1,12 +1,12 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
-  transpilePackages: ['@zama-ai/service-status-monitor'],
+  transpilePackages: ['zama-service-status-monitor'],
   // Enable watching changes in workspace packages
   webpack: (config) => {
     config.watchOptions = {
       ...config.watchOptions,
-      ignored: /node_modules([\\]+|\/)+(?!@zama-ai)/,
+      ignored: /node_modules([\\]+|\/)+(?!zama-service-status-monitor)/,
     };
     return config;
   },